refactor(ai): move the Anthropic tool-reference pass into its own module
The tool-reference integrity pass (demoteUnavailableToolReferences and its
helpers) was a fork-only block inside the 2.7k-line anthropic-messages.ts.
Move it to anthropic-tool-references.ts next to anthropic-tool-pairs.ts, and
split its test suite into a shared harness plus one file per behavior cluster
(client tool_reference/tool_use integrity, native tool-search replay).
No behavior change.
